Refactored API related routes into a separate module api.js
